+++ build.gradle
@@ -24,7 +24,6 @@ dependencies {
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'io.rest-assured:rest-assured:5.3.1'
-    runtimeOnly 'com.h2database:h2'
 }
 
 test {
