+++ src/main/java/roomescape/dao/ReservationDAO.java
@@ -1,15 +1,19 @@
-package roomescape.model;
+package roomescape.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.model.Reservation;
 
+import java.sql.PreparedStatement;
 import java.util.List;
 
 @Repository
-public class QueryingDAO {
+public class ReservationDAO {
     private final JdbcTemplate jdbcTemplate;
 
-    public QueryingDAO(JdbcTemplate jdbcTemplate) {
+    public ReservationDAO(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
@@ -40,5 +44,24 @@ public class QueryingDAO {
                     );
                     return reservation;
                 }, id);
+    }
+
+    public Long insertWithKeyHolder(Reservation reservation) {
+        String sql = "insert into reservation (name,date,time) values(?,?,?)";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
+            ps.setString(1, reservation.getName());
+            ps.setString(2, reservation.getDate());
+            ps.setString(3, reservation.getTime());
+            return ps;
+        }, keyHolder);
+
+        return keyHolder.getKey().longValue();
+    }
+
+    public void delete(Long id) {
+        String sql = "delete from reservation where id = ?";
+        jdbcTemplate.update(sql, id);
     }
 }
